version LtenRMU 4.9: done

+++ src/evals/metrics/utils.py
@@ -95,8 +95,10 @@ def evaluate_probability(model, batch):
     avg_losses = losses / num_token_gt
     normalized_probs = torch.exp(-avg_losses)
 
-    avg_losses = avg_losses.cpu().numpy().tolist()
-    normalized_probs = normalized_probs.cpu().numpy().tolist()
+    # transformers >=5 keeps logits in bf16 unless bf16_full_eval is off;
+    # numpy can't convert bf16 directly, so cast to float32 first.
+    avg_losses = avg_losses.float().cpu().numpy().tolist()
+    normalized_probs = normalized_probs.float().cpu().numpy().tolist()
     return [
         {"prob": prob, "avg_loss": avg_loss}
         for prob, avg_loss in zip(normalized_probs, avg_losses)
